+++ ShipmentGeek/Properties/AssemblyInfo.cs
@@ -6,7 +6,7 @@ using System.Runtime.InteropServices;
 // set of attributes. Change these attribute values to modify the information
 // associated with an assembly.
 [assembly: AssemblyTitle("Shipment Geek")]
-[assembly: AssemblyDescription("Keep track of incoming and outgoing shipments")]
+[assembly: AssemblyDescription("Keep track of incoming and outgoing shipments.")]
 [assembly: AssemblyConfiguration("")]
 [assembly: AssemblyCompany("")]
 [assembly: AssemblyProduct("ShipmentGeek")]