+++ ShipmentGeek/frmAbout.cs
@@ -1,0 +1,102 @@
+using System;
+using System.Collections.Generic;
+using System.ComponentModel;
+using System.Drawing;
+using System.Windows.Forms;
+using System.Reflection;
+
+namespace ShipmentGeek
+{
+    partial class frmAbout : Form
+    {
+        public frmAbout(string version, string library, string developedBy)
+        {
+            InitializeComponent();
+
+            //  Initialize the AboutBox to display the product information from the assembly information.
+            //  Change assembly information settings for your application through either:
+            //  - Project->Properties->Application->Assembly Information
+            //  - AssemblyInfo.cs
+            this.Text = String.Format("About {0}", AssemblyTitle);
+            this.labelProductName.Text = String.Format("{0} v.{1}", AssemblyTitle, version);
+            this.labelLibrary.Text = String.Format("Library v.{0}", library);
+            this.labelCopyright.Text = AssemblyCopyright;
+            this.labelCompanyName.Text = AssemblyCompany;
+            this.labelDevelopedBy.Text = developedBy;
+            this.textBoxDescription.Text = AssemblyDescription;
+
+#if DEBUG
+            this.labelConfiguration.Text = "Debug";
+            this.labelConfiguration.Image = Properties.Resources.bug_icon;
+#else
+            this.labelConfiguration.Text = "Release";
+            this.labelConfiguration.Image = Properties.Resources.shipment_16;
+#endif
+        }
+
+        #region Assembly Attribute Accessors
+
+        public string AssemblyTitle
+        {
+            get
+            {
+                // Get all Title attributes on this assembly
+                object[] attributes = Assembly.GetExecutingAssembly().GetCustomAttributes(typeof(AssemblyTitleAttribute), false);
+                // If there is at least one Title attribute
+                if (attributes.Length > 0)
+                {
+                    // Select the first one
+                    AssemblyTitleAttribute titleAttribute = (AssemblyTitleAttribute)attributes[0];
+                    // If it is not an empty string, return it
+                    if (titleAttribute.Title != "")
+                        return titleAttribute.Title;
+                }
+                // If there was no Title attribute, or if the Title attribute was the empty string, return the .exe name
+                return System.IO.Path.GetFileNameWithoutExtension(Assembly.GetExecutingAssembly().CodeBase);
+            }
+        }
+
+        public string AssemblyDescription
+        {
+            get
+            {
+                // Get all Description attributes on this assembly
+                object[] attributes = Assembly.GetExecutingAssembly().GetCustomAttributes(typeof(AssemblyDescriptionAttribute), false);
+                // If there aren't any Description attributes, return an empty string
+                if (attributes.Length == 0)
+                    return "";
+                // If there is a Description attribute, return its value
+                return ((AssemblyDescriptionAttribute)attributes[0]).Description;
+            }
+        }
+
+        public string AssemblyCopyright
+        {
+            get
+            {
+                // Get all Copyright attributes on this assembly
+                object[] attributes = Assembly.GetExecutingAssembly().GetCustomAttributes(typeof(AssemblyCopyrightAttribute), false);
+                // If there aren't any Copyright attributes, return an empty string
+                if (attributes.Length == 0)
+                    return "";
+                // If there is a Copyright attribute, return its value
+                return ((AssemblyCopyrightAttribute)attributes[0]).Copyright;
+            }
+        }
+
+        public string AssemblyCompany
+        {
+            get
+            {
+                // Get all Company attributes on this assembly
+                object[] attributes = Assembly.GetExecutingAssembly().GetCustomAttributes(typeof(AssemblyCompanyAttribute), false);
+                // If there aren't any Company attributes, return an empty string
+                if (attributes.Length == 0)
+                    return "";
+                // If there is a Company attribute, return its value
+                return ((AssemblyCompanyAttribute)attributes[0]).Company;
+            }
+        }
+        #endregion
+    }
+}