+++ ShipmentGeek/frmMain.Designer.cs
@@ -68,6 +68,7 @@
             this.statStrip = new System.Windows.Forms.StatusStrip();
             this.srpSelected = new System.Windows.Forms.ToolStripStatusLabel();
             this.srpLoadSave = new System.Windows.Forms.ToolStripStatusLabel();
+            this.srpSearch = new System.Windows.Forms.ToolStripStatusLabel();
             this.grpIncoming = new System.Windows.Forms.GroupBox();
             this.lstIncoming = new System.Windows.Forms.ListView();
             this.clmInID = ((System.Windows.Forms.ColumnHeader)(new System.Windows.Forms.ColumnHeader()));
@@ -90,6 +91,13 @@
             this.splitDetails = new System.Windows.Forms.SplitContainer();
             this.cmdNew = new System.Windows.Forms.Button();
             this.toolTip = new System.Windows.Forms.ToolTip(this.components);
+            this.mnuStrip = new System.Windows.Forms.MenuStrip();
+            this.mnuFile = new System.Windows.Forms.ToolStripMenuItem();
+            this.mnuExit = new System.Windows.Forms.ToolStripMenuItem();
+            this.mnuShipment = new System.Windows.Forms.ToolStripMenuItem();
+            this.mnuShipmentSearch = new System.Windows.Forms.ToolStripMenuItem();
+            this.mnuHelp = new System.Windows.Forms.ToolStripMenuItem();
+            this.mnuAbout = new System.Windows.Forms.ToolStripMenuItem();
             this.grpDetails.SuspendLayout();
             this.tableLayoutPanel1.SuspendLayout();
             this.panType.SuspendLayout();
@@ -110,6 +118,7 @@
             this.splitDetails.Panel1.SuspendLayout();
             this.splitDetails.Panel2.SuspendLayout();
             this.splitDetails.SuspendLayout();
+            this.mnuStrip.SuspendLayout();
             this.SuspendLayout();
             // 
             // cmdSave
@@ -118,10 +127,10 @@
             this.cmdSave.Enabled = false;
             this.cmdSave.Image = global::ShipmentGeek.Properties.Resources.save;
             this.cmdSave.ImageAlign = System.Drawing.ContentAlignment.MiddleLeft;
-            this.cmdSave.Location = new System.Drawing.Point(93, 502);
+            this.cmdSave.Location = new System.Drawing.Point(93, 529);
             this.cmdSave.Name = "cmdSave";
             this.cmdSave.Size = new System.Drawing.Size(75, 23);
-            this.cmdSave.TabIndex = 2;
+            this.cmdSave.TabIndex = 3;
             this.cmdSave.Text = "Save";
             this.toolTip.SetToolTip(this.cmdSave, "Save selected shipment");
             this.cmdSave.UseVisualStyleBackColor = true;
@@ -135,7 +144,7 @@
             this.grpDetails.Controls.Add(this.tableLayoutPanel1);
             this.grpDetails.Location = new System.Drawing.Point(3, 3);
             this.grpDetails.Name = "grpDetails";
-            this.grpDetails.Size = new System.Drawing.Size(355, 289);
+            this.grpDetails.Size = new System.Drawing.Size(355, 295);
             this.grpDetails.TabIndex = 0;
             this.grpDetails.TabStop = false;
             this.grpDetails.Text = "Shipment Details";
@@ -178,7 +187,7 @@
             this.tableLayoutPanel1.RowStyles.Add(new System.Windows.Forms.RowStyle(System.Windows.Forms.SizeType.Absolute, 26F));
             this.tableLayoutPanel1.RowStyles.Add(new System.Windows.Forms.RowStyle(System.Windows.Forms.SizeType.Percent, 100F));
             this.tableLayoutPanel1.RowStyles.Add(new System.Windows.Forms.RowStyle(System.Windows.Forms.SizeType.Absolute, 26F));
-            this.tableLayoutPanel1.Size = new System.Drawing.Size(343, 264);
+            this.tableLayoutPanel1.Size = new System.Drawing.Size(343, 270);
             this.tableLayoutPanel1.TabIndex = 0;
             // 
             // lblType
@@ -224,7 +233,7 @@
             this.txtComment.Multiline = true;
             this.txtComment.Name = "txtComment";
             this.txtComment.ScrollBars = System.Windows.Forms.ScrollBars.Vertical;
-            this.txtComment.Size = new System.Drawing.Size(252, 50);
+            this.txtComment.Size = new System.Drawing.Size(252, 56);
             this.txtComment.TabIndex = 8;
             // 
             // panType
@@ -270,7 +279,7 @@
             // lblReceived
             // 
             this.lblReceived.AutoSize = true;
-            this.lblReceived.Location = new System.Drawing.Point(3, 238);
+            this.lblReceived.Location = new System.Drawing.Point(3, 244);
             this.lblReceived.Name = "lblReceived";
             this.lblReceived.Size = new System.Drawing.Size(37, 13);
             this.lblReceived.TabIndex = 9;
@@ -340,7 +349,7 @@
             // 
             this.panStatus.Controls.Add(this.chkReceived);
             this.panStatus.Controls.Add(this.chkMissing);
-            this.panStatus.Location = new System.Drawing.Point(88, 241);
+            this.panStatus.Location = new System.Drawing.Point(88, 247);
             this.panStatus.Name = "panStatus";
             this.panStatus.Size = new System.Drawing.Size(252, 19);
             this.panStatus.TabIndex = 9;
@@ -445,7 +454,7 @@
             this.cmdItemSave.Location = new System.Drawing.Point(319, 16);
             this.cmdItemSave.Name = "cmdItemSave";
             this.cmdItemSave.Size = new System.Drawing.Size(30, 23);
-            this.cmdItemSave.TabIndex = 2;
+            this.cmdItemSave.TabIndex = 3;
             this.toolTip.SetToolTip(this.cmdItemSave, "Save selected shipment item");
             this.cmdItemSave.UseVisualStyleBackColor = true;
             this.cmdItemSave.Click += new System.EventHandler(this.cmdItemSave_Click);
@@ -471,7 +480,7 @@
             this.lstItems.Location = new System.Drawing.Point(6, 45);
             this.lstItems.MultiSelect = false;
             this.lstItems.Name = "lstItems";
-            this.lstItems.Size = new System.Drawing.Size(343, 162);
+            this.lstItems.Size = new System.Drawing.Size(343, 168);
             this.lstItems.TabIndex = 4;
             this.lstItems.UseCompatibleStateImageBehavior = false;
             this.lstItems.View = System.Windows.Forms.View.Details;
@@ -493,8 +502,9 @@
             // 
             this.statStrip.Items.AddRange(new System.Windows.Forms.ToolStripItem[] {
             this.srpSelected,
-            this.srpLoadSave});
-            this.statStrip.Location = new System.Drawing.Point(0, 531);
+            this.srpLoadSave,
+            this.srpSearch});
+            this.statStrip.Location = new System.Drawing.Point(0, 558);
             this.statStrip.Name = "statStrip";
             this.statStrip.Size = new System.Drawing.Size(774, 22);
             this.statStrip.TabIndex = 4;
@@ -518,6 +528,14 @@
             this.srpLoadSave.Text = "LoadSave";
             this.srpLoadSave.Visible = false;
             // 
+            // srpSearch
+            // 
+            this.srpSearch.Image = global::ShipmentGeek.Properties.Resources.search;
+            this.srpSearch.Name = "srpSearch";
+            this.srpSearch.Size = new System.Drawing.Size(58, 17);
+            this.srpSearch.Text = "Search";
+            this.srpSearch.Visible = false;
+            // 
             // grpIncoming
             // 
             this.grpIncoming.Anchor = ((System.Windows.Forms.AnchorStyles)((((System.Windows.Forms.AnchorStyles.Top | System.Windows.Forms.AnchorStyles.Bottom) 
@@ -526,7 +544,7 @@
             this.grpIncoming.Controls.Add(this.lstIncoming);
             this.grpIncoming.Location = new System.Drawing.Point(3, 3);
             this.grpIncoming.Name = "grpIncoming";
-            this.grpIncoming.Size = new System.Drawing.Size(377, 235);
+            this.grpIncoming.Size = new System.Drawing.Size(377, 241);
             this.grpIncoming.TabIndex = 0;
             this.grpIncoming.TabStop = false;
             this.grpIncoming.Text = "Incoming Shipments";
@@ -548,7 +566,7 @@
             this.lstIncoming.Location = new System.Drawing.Point(6, 19);
             this.lstIncoming.MultiSelect = false;
             this.lstIncoming.Name = "lstIncoming";
-            this.lstIncoming.Size = new System.Drawing.Size(365, 208);
+            this.lstIncoming.Size = new System.Drawing.Size(365, 214);
             this.lstIncoming.TabIndex = 0;
             this.lstIncoming.UseCompatibleStateImageBehavior = false;
             this.lstIncoming.View = System.Windows.Forms.View.Details;
@@ -590,7 +608,7 @@
             this.grpOutgoing.Controls.Add(this.lstOutgoing);
             this.grpOutgoing.Location = new System.Drawing.Point(3, 3);
             this.grpOutgoing.Name = "grpOutgoing";
-            this.grpOutgoing.Size = new System.Drawing.Size(377, 231);
+            this.grpOutgoing.Size = new System.Drawing.Size(377, 237);
             this.grpOutgoing.TabIndex = 1;
             this.grpOutgoing.TabStop = false;
             this.grpOutgoing.Text = "Outgoing Shipments";
@@ -612,7 +630,7 @@
             this.lstOutgoing.Location = new System.Drawing.Point(6, 19);
             this.lstOutgoing.MultiSelect = false;
             this.lstOutgoing.Name = "lstOutgoing";
-            this.lstOutgoing.Size = new System.Drawing.Size(365, 206);
+            this.lstOutgoing.Size = new System.Drawing.Size(365, 212);
             this.lstOutgoing.TabIndex = 0;
             this.lstOutgoing.UseCompatibleStateImageBehavior = false;
             this.lstOutgoing.View = System.Windows.Forms.View.Details;
@@ -651,7 +669,7 @@
             this.chkShowAll.Anchor = ((System.Windows.Forms.AnchorStyles)((System.Windows.Forms.AnchorStyles.Bottom | System.Windows.Forms.AnchorStyles.Right)));
             this.chkShowAll.AutoSize = true;
             this.chkShowAll.CheckAlign = System.Drawing.ContentAlignment.MiddleRight;
-            this.chkShowAll.Location = new System.Drawing.Point(277, 506);
+            this.chkShowAll.Location = new System.Drawing.Point(277, 533);
             this.chkShowAll.Name = "chkShowAll";
             this.chkShowAll.Size = new System.Drawing.Size(118, 17);
             this.chkShowAll.TabIndex = 4;
@@ -676,7 +694,7 @@
             this.grpItems.Enabled = false;
             this.grpItems.Location = new System.Drawing.Point(3, 3);
             this.grpItems.Name = "grpItems";
-            this.grpItems.Size = new System.Drawing.Size(355, 213);
+            this.grpItems.Size = new System.Drawing.Size(355, 219);
             this.grpItems.TabIndex = 1;
             this.grpItems.TabStop = false;
             this.grpItems.Text = "Shipment Items";
@@ -687,7 +705,7 @@
             this.cmdItemAdd.Location = new System.Drawing.Point(283, 16);
             this.cmdItemAdd.Name = "cmdItemAdd";
             this.cmdItemAdd.Size = new System.Drawing.Size(30, 23);
-            this.cmdItemAdd.TabIndex = 6;
+            this.cmdItemAdd.TabIndex = 2;
             this.toolTip.SetToolTip(this.cmdItemAdd, "Add new shipment item");
             this.cmdItemAdd.UseVisualStyleBackColor = true;
             this.cmdItemAdd.Click += new System.EventHandler(this.cmdItemSave_Click);
@@ -697,7 +715,7 @@
             this.splitShipments.Anchor = ((System.Windows.Forms.AnchorStyles)((((System.Windows.Forms.AnchorStyles.Top | System.Windows.Forms.AnchorStyles.Bottom) 
             | System.Windows.Forms.AnchorStyles.Left) 
             | System.Windows.Forms.AnchorStyles.Right)));
-            this.splitShipments.Location = new System.Drawing.Point(12, 12);
+            this.splitShipments.Location = new System.Drawing.Point(12, 27);
             this.splitShipments.Name = "splitShipments";
             this.splitShipments.Orientation = System.Windows.Forms.Orientation.Horizontal;
             // 
@@ -710,15 +728,15 @@
             // 
             this.splitShipments.Panel2.Controls.Add(this.grpOutgoing);
             this.splitShipments.Panel2MinSize = 100;
-            this.splitShipments.Size = new System.Drawing.Size(383, 482);
-            this.splitShipments.SplitterDistance = 241;
+            this.splitShipments.Size = new System.Drawing.Size(383, 494);
+            this.splitShipments.SplitterDistance = 247;
             this.splitShipments.TabIndex = 0;
             // 
             // splitDetails
             // 
             this.splitDetails.Anchor = ((System.Windows.Forms.AnchorStyles)(((System.Windows.Forms.AnchorStyles.Top | System.Windows.Forms.AnchorStyles.Bottom) 
             | System.Windows.Forms.AnchorStyles.Right)));
-            this.splitDetails.Location = new System.Drawing.Point(401, 12);
+            this.splitDetails.Location = new System.Drawing.Point(401, 27);
             this.splitDetails.Name = "splitDetails";
             this.splitDetails.Orientation = System.Windows.Forms.Orientation.Horizontal;
             // 
@@ -731,8 +749,8 @@
             // 
             this.splitDetails.Panel2.Controls.Add(this.grpItems);
             this.splitDetails.Panel2MinSize = 125;
-            this.splitDetails.Size = new System.Drawing.Size(361, 516);
-            this.splitDetails.SplitterDistance = 293;
+            this.splitDetails.Size = new System.Drawing.Size(361, 528);
+            this.splitDetails.SplitterDistance = 299;
             this.splitDetails.TabIndex = 1;
             // 
             // cmdNew
@@ -740,27 +758,92 @@
             this.cmdNew.Anchor = ((System.Windows.Forms.AnchorStyles)((System.Windows.Forms.AnchorStyles.Bottom | System.Windows.Forms.AnchorStyles.Left)));
             this.cmdNew.Image = global::ShipmentGeek.Properties.Resources.add;
             this.cmdNew.ImageAlign = System.Drawing.ContentAlignment.MiddleLeft;
-            this.cmdNew.Location = new System.Drawing.Point(12, 502);
+            this.cmdNew.Location = new System.Drawing.Point(12, 529);
             this.cmdNew.Name = "cmdNew";
             this.cmdNew.Size = new System.Drawing.Size(75, 23);
-            this.cmdNew.TabIndex = 5;
+            this.cmdNew.TabIndex = 2;
             this.cmdNew.Text = "New";
             this.toolTip.SetToolTip(this.cmdNew, "Add new shipment");
             this.cmdNew.UseVisualStyleBackColor = true;
             this.cmdNew.Click += new System.EventHandler(this.cmdSave_Click);
             // 
+            // mnuStrip
+            // 
+            this.mnuStrip.Items.AddRange(new System.Windows.Forms.ToolStripItem[] {
+            this.mnuFile,
+            this.mnuShipment,
+            this.mnuHelp});
+            this.mnuStrip.Location = new System.Drawing.Point(0, 0);
+            this.mnuStrip.Name = "mnuStrip";
+            this.mnuStrip.Size = new System.Drawing.Size(774, 24);
+            this.mnuStrip.TabIndex = 6;
+            this.mnuStrip.Text = "menuStrip1";
+            // 
+            // mnuFile
+            // 
+            this.mnuFile.DropDownItems.AddRange(new System.Windows.Forms.ToolStripItem[] {
+            this.mnuExit});
+            this.mnuFile.Name = "mnuFile";
+            this.mnuFile.Size = new System.Drawing.Size(37, 20);
+            this.mnuFile.Text = "File";
+            // 
+            // mnuExit
+            // 
+            this.mnuExit.Image = global::ShipmentGeek.Properties.Resources.exit;
+            this.mnuExit.Name = "mnuExit";
+            this.mnuExit.ShortcutKeys = ((System.Windows.Forms.Keys)((System.Windows.Forms.Keys.Alt | System.Windows.Forms.Keys.F4)));
+            this.mnuExit.Size = new System.Drawing.Size(134, 22);
+            this.mnuExit.Text = "Exit";
+            this.mnuExit.Click += new System.EventHandler(this.mnuExit_Click);
+            // 
+            // mnuShipment
+            // 
+            this.mnuShipment.DropDownItems.AddRange(new System.Windows.Forms.ToolStripItem[] {
+            this.mnuShipmentSearch});
+            this.mnuShipment.Name = "mnuShipment";
+            this.mnuShipment.Size = new System.Drawing.Size(70, 20);
+            this.mnuShipment.Text = "Shipment";
+            // 
+            // mnuShipmentSearch
+            // 
+            this.mnuShipmentSearch.Image = global::ShipmentGeek.Properties.Resources.search;
+            this.mnuShipmentSearch.Name = "mnuShipmentSearch";
+            this.mnuShipmentSearch.ShortcutKeys = ((System.Windows.Forms.Keys)((System.Windows.Forms.Keys.Control | System.Windows.Forms.Keys.F)));
+            this.mnuShipmentSearch.Size = new System.Drawing.Size(137, 22);
+            this.mnuShipmentSearch.Text = "Find";
+            this.mnuShipmentSearch.Click += new System.EventHandler(this.mnuShipmentSearch_Click);
+            // 
+            // mnuHelp
+            // 
+            this.mnuHelp.DropDownItems.AddRange(new System.Windows.Forms.ToolStripItem[] {
+            this.mnuAbout});
+            this.mnuHelp.Name = "mnuHelp";
+            this.mnuHelp.Size = new System.Drawing.Size(44, 20);
+            this.mnuHelp.Text = "Help";
+            // 
+            // mnuAbout
+            // 
+            this.mnuAbout.Image = global::ShipmentGeek.Properties.Resources.about;
+            this.mnuAbout.Name = "mnuAbout";
+            this.mnuAbout.ShortcutKeys = ((System.Windows.Forms.Keys)((System.Windows.Forms.Keys.Shift | System.Windows.Forms.Keys.F1)));
+            this.mnuAbout.Size = new System.Drawing.Size(241, 22);
+            this.mnuAbout.Text = "About Shipment Geek";
+            this.mnuAbout.Click += new System.EventHandler(this.mnuAbout_Click);
+            // 
             // frmMain
             // 
             this.AutoScaleDimensions = new System.Drawing.SizeF(6F, 13F);
             this.AutoScaleMode = System.Windows.Forms.AutoScaleMode.Font;
-            this.ClientSize = new System.Drawing.Size(774, 553);
+            this.ClientSize = new System.Drawing.Size(774, 580);
             this.Controls.Add(this.cmdNew);
             this.Controls.Add(this.splitDetails);
             this.Controls.Add(this.splitShipments);
             this.Controls.Add(this.chkShowAll);
             this.Controls.Add(this.statStrip);
+            this.Controls.Add(this.mnuStrip);
             this.Controls.Add(this.cmdSave);
             this.Icon = ((System.Drawing.Icon)(resources.GetObject("$this.Icon")));
+            this.MainMenuStrip = this.mnuStrip;
             this.Name = "frmMain";
             this.Text = "Shipment Geek";
             this.Load += new System.EventHandler(this.frmMain_Load);
@@ -791,6 +874,8 @@
             this.splitDetails.Panel2.ResumeLayout(false);
             ((System.ComponentModel.ISupportInitialize)(this.splitDetails)).EndInit();
             this.splitDetails.ResumeLayout(false);
+            this.mnuStrip.ResumeLayout(false);
+            this.mnuStrip.PerformLayout();
             this.ResumeLayout(false);
             this.PerformLayout();
 
@@ -858,6 +943,14 @@
         private System.Windows.Forms.Button cmdItemAdd;
         private System.Windows.Forms.ToolTip toolTip;
         private System.Windows.Forms.ToolStripStatusLabel srpLoadSave;
+        private System.Windows.Forms.ToolStripStatusLabel srpSearch;
+        private System.Windows.Forms.MenuStrip mnuStrip;
+        private System.Windows.Forms.ToolStripMenuItem mnuFile;
+        private System.Windows.Forms.ToolStripMenuItem mnuHelp;
+        private System.Windows.Forms.ToolStripMenuItem mnuExit;
+        private System.Windows.Forms.ToolStripMenuItem mnuAbout;
+        private System.Windows.Forms.ToolStripMenuItem mnuShipment;
+        private System.Windows.Forms.ToolStripMenuItem mnuShipmentSearch;
 
     }
 }